Remove pride, change to HolidayProvider
