+++ apps/docs/app/root/layout/SiteSettings.tsx
@@ -12,7 +12,6 @@ import {
   Switch,
   Text,
   useColorMode,
-  TogglePride,
 } from "@vygruppen/spor-react";
 import { BrandSwitcher } from "~/features/brand-switcher/BrandSwitcher";
 
@@ -54,7 +53,6 @@ export const SiteSettings = ({ showLabel }: SiteSettingsProps) => {
                 defaultChecked={colorMode === "dark"}
               />
             </FormControl>
-            <TogglePride label="Make it pride" />
           </Stack>
         </Flex>
       </CardSelect>
