+++ packages/spor-react/src/loader/ColorInlineLoader.tsx
@@ -3,8 +3,6 @@ import { inlineLoaderColorData } from "@vygruppen/spor-loader";
 import React from "react";
 import { ClientOnly } from "./ClientOnly";
 import Lottie from "./Lottie";
-import { inlineLoaderColorPrideData } from "@vygruppen/spor-loader";
-import { usePride } from "../pride";
 
 export type ColorInlineLoaderProps = Exclude<BoxProps, "children">;
 /**
@@ -17,18 +15,11 @@ export const ColorInlineLoader = ({
   maxWidth,
   ...props
 }: ColorInlineLoaderProps) => {
-  const { isPride } = usePride();
   return (
     <Center {...props}>
       <Box width={width} maxWidth={maxWidth}>
         <ClientOnly>
-          {() => (
-            <Lottie
-              animationData={
-                isPride ? inlineLoaderColorPrideData : inlineLoaderColorData
-              }
-            />
-          )}
+          {() => <Lottie animationData={inlineLoaderColorData} />}
         </ClientOnly>
       </Box>
     </Center>
