+++ packages/spor-react/src/loader/ColorSpinner.tsx
@@ -1,12 +1,8 @@
 import { Box, BoxProps, Center } from "@chakra-ui/react";
-import {
-  spinnerColorData,
-  spinnerColorPrideData,
-} from "@vygruppen/spor-loader";
+import { spinnerColorData } from "@vygruppen/spor-loader";
 import React from "react";
 import { ClientOnly } from "./ClientOnly";
 import Lottie from "./Lottie";
-import { usePride } from "../pride/PrideProvider";
 
 export type SpinnerProps = BoxProps;
 export type ColorSpinnerProps = SpinnerProps;
@@ -32,16 +28,11 @@ export const ColorSpinner = ({
 
   ...props
 }: SpinnerProps) => {
-  const { isPride } = usePride();
   return (
     <Center flexDirection="column" {...props}>
       <Box width={width} maxWidth={maxWidth}>
         <ClientOnly>
-          {() => (
-            <Lottie
-              animationData={isPride ? spinnerColorPrideData : spinnerColorData}
-            />
-          )}
+          {() => <Lottie animationData={spinnerColorData} />}
         </ClientOnly>
       </Box>
       {children && (
