+++ packages/spor-react/src/loader/ContentLoader.tsx
@@ -1,12 +1,8 @@
 import { Box, BoxProps } from "@chakra-ui/react";
-import {
-  contentLoaderData,
-  contentLoaderPrideData,
-} from "@vygruppen/spor-loader";
+import { contentLoaderData } from "@vygruppen/spor-loader";
 import React from "react";
 import { ClientOnly } from "./ClientOnly";
 import Lottie from "./Lottie";
-import { usePride } from "../pride/PrideProvider";
 
 export type ContentLoaderProps = BoxProps;
 /**
@@ -14,18 +10,11 @@ export type ContentLoaderProps = BoxProps;
  * It should mostly be used for
  */
 export const ContentLoader = ({ children, ...props }: ContentLoaderProps) => {
-  const { isPride } = usePride();
   return (
     <Box {...props}>
       <Box maxWidth="140px" marginX="auto">
         <ClientOnly>
-          {() => (
-            <Lottie
-              animationData={
-                isPride ? contentLoaderPrideData : contentLoaderData
-              }
-            />
-          )}
+          {() => <Lottie animationData={contentLoaderData} />}
         </ClientOnly>
       </Box>
       {children && (
