+++ packages/spor-react/src/loader/DarkInlineLoader.tsx
@@ -1,12 +1,8 @@
 import { Box, BoxProps, Center } from "@chakra-ui/react";
-import {
-  inlineLoaderColorPrideData,
-  inlineLoaderDarkData,
-} from "@vygruppen/spor-loader";
+import { inlineLoaderDarkData } from "@vygruppen/spor-loader";
 import React from "react";
 import { ClientOnly } from "./ClientOnly";
 import Lottie from "./Lottie";
-import { usePride } from "../pride/PrideProvider";
 
 export type DarkInlineLoaderProps = Exclude<BoxProps, "children">;
 /**
@@ -17,18 +13,11 @@ export const DarkInlineLoader = ({
   maxWidth,
   ...props
 }: DarkInlineLoaderProps) => {
-  const { isPride } = usePride();
   return (
     <Center {...props}>
       <Box width={width} maxWidth={maxWidth}>
         <ClientOnly>
-          {() => (
-            <Lottie
-              animationData={
-                isPride ? inlineLoaderColorPrideData : inlineLoaderDarkData
-              }
-            />
-          )}
+          {() => <Lottie animationData={inlineLoaderDarkData} />}
         </ClientOnly>
       </Box>
     </Center>
