+++ packages/spor-react/src/loader/LightFullScreenLoader.tsx
@@ -3,8 +3,6 @@ import { fullScreenLoaderBlackData } from "@vygruppen/spor-loader";
 import React from "react";
 import { ClientOnly } from "./ClientOnly";
 import Lottie from "./Lottie";
-import { usePride } from "../pride";
-import { vyLogoPrideData } from "@vygruppen/spor-loader";
 
 type LightFullScreenLoaderProps = BoxProps;
 
@@ -13,18 +11,11 @@ export const LightFullScreenLoader = ({
   maxWidth,
   ...props
 }: LightFullScreenLoaderProps) => {
-  const { isPride } = usePride();
   return (
     <Center height="100%" background="white" {...props}>
       <Box width={width} maxWidth={maxWidth}>
         <ClientOnly>
-          {() => (
-            <Lottie
-              animationData={
-                isPride ? vyLogoPrideData : fullScreenLoaderBlackData
-              }
-            />
-          )}
+          {() => <Lottie animationData={fullScreenLoaderBlackData} />}
         </ClientOnly>
       </Box>
     </Center>
