+++ packages/spor-react/src/loader/LightInlineLoader.tsx
@@ -1,12 +1,8 @@
 import { Box, BoxProps, Center } from "@chakra-ui/react";
-import {
-  inlineLoaderColorPrideData,
-  inlineLoaderLightData,
-} from "@vygruppen/spor-loader";
+import { inlineLoaderLightData } from "@vygruppen/spor-loader";
 import React from "react";
 import { ClientOnly } from "./ClientOnly";
 import Lottie from "./Lottie";
-import { usePride } from "../pride/PrideProvider";
 
 export type LightInlineLoaderProps = Exclude<BoxProps, "children">;
 /**
@@ -17,18 +13,11 @@ export const LightInlineLoader = ({
   maxWidth,
   ...props
 }: LightInlineLoaderProps) => {
-  const { isPride } = usePride();
   return (
     <Center {...props}>
       <Box width={width} maxWidth={maxWidth}>
         <ClientOnly>
-          {() => (
-            <Lottie
-              animationData={
-                isPride ? inlineLoaderColorPrideData : inlineLoaderLightData
-              }
-            />
-          )}
+          {() => <Lottie animationData={inlineLoaderLightData} />}
         </ClientOnly>
       </Box>
     </Center>
