+++ packages/spor-react/src/logo/index.tsx
@@ -1,4 +1,3 @@
 export * from "./VyLogo";
 export * from "./VyLogoPride";
-export * from "./VyLogoDefault";
 export * from "./CargonetLogo";
