--- packages/spor-react/src/pride/index.tsx
@@ -1,2 +1,0 @@
-export * from "./PrideProvider";
-export * from "./TogglePride";
